scidvspc my snap initial commit to git
This is a attempt at snappification of scidvspc, so that it can be
used in different versions of ubuntu, without worrying about what
dependencies are available in it or not.

This is the Alphabet A in the steps towards snappification.

Have added a basic skeleton of a snapcraft.yaml.

NOTE: commit to git from backup files