TclTkInSnap:P?: part sequencing, nil plugin
Instead of copying / overwriting a dest file, edit it in place.

As we dont require any source files, so change to nil plugin from
dump.

NOTE: Also now after so much time, I am not sure why I had swapped
source and organize->dest wrt dump plugin based part, at that time.
Was it a mistake or was I trying to see if dump can work in reverse
or ...