scidvspc snap package initial submission similar
This is snappification of scidvspc, so that it can be used in
different versions of ubuntu, without worrying about what
dependencies are available in it or not.

Have tried to keep the set of permissions requested to a minimum.

A similar version was uploaded to ubuntu snap repository earlier.

However given that it is specified has needing unity7 interface,
it appears that it also needs a desktop file, so the submission
had got rejected.