A look after 2 years - new snapcraft base tag
Have updated the version to indicate todays date.

Snapcraft new version now seems to be using virtual machines based
around the ubuntu multipass package for building the snaps, instead
of the previous lxd and the cleanbuild built on top of it.

In order to work with this new mechanism, I have added the base tag,
which is needed, to snapcraft yaml file, which if I understand correctly
basically indicates what is the base set of packages / environment for
the vm used for building.