Dont hide mkdir on existing prg src dir
This will ensure that I can know when I have run the snapcraft
build more than once on same vm-build instance.

Running build more than once can mess with path fix logic, for now,
so this acts as a coarse grained alert to it.

Also make a reference to the github repository.