Sounds: Set the soundFolder by default
As the base path /snap/ is by default not browsable by the user from
with in the snap, while the specific snap package's subfolder within
it is browsable. So if the user tries to set the sounds folder by
trying to browse to /snap and then select the subfolders from there,
it wont work. While if they directly enter /snap/scidvspc-hkvc, then
they can browse the remaining subfolders AND OR one can even directly
enter the full path.

So I am setting the program provided sounds folder by default now,
so that user can just go about using the sound, if they want to,
without worrying about trying to find where the sounds are or how
to set it, given the blackhole of the /snap/.