core20 Explicitly install libsnack-alsa
The snack in core20 was using libsnack-oss by default, which inturn
was giving the error wrt /dev/mixer and /dev/sound/dsp ...

So now I explicitly specify that libsnack-alsa should be used in
the snap runtime, this fixes this problem with missing devices.

However I notice that with core20 now for some reason snack is
playing only every 2nd sound so white moves are played but black
moves seem to be silent??? I vaguely remember reading some similar
issue on internet wrt snack and scidvspc, when I was looking into
some other issue few days back. And it was talking about snack
being buggy or so and needing a maintainer, need to look at this
later.